// src/main/webapp/src/app/rest/model/changeLogEvent.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChangeLogEvent { 
    timestamp?: string;
    seq?: number;
    eventType: string;
}