// src/main/webapp/src/app/rest/model/githubPRChange.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChangeLogEvent } from './changeLogEvent';


export interface GithubPRChange extends ChangeLogEvent { 
    number?: number;
    changeType?: string;
}