// src/main/webapp/src/app/rest/model/jiraIssueChange.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChangeLogEvent } from './changeLogEvent';


export interface JiraIssueChange extends ChangeLogEvent { 
    issueKey?: string;
    changeType?: string;
}