// src/main/webapp/src/app/rest/model/mailingListChange.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChangeLogEvent } from './changeLogEvent';


export interface MailingListChange extends ChangeLogEvent { 
    messageId?: string;
    subject?: string;
    changeType?: string;
}